Add unknown event.

+++ lib/blocs/bloc_jwt.dart
@@ -4,6 +4,7 @@ import 'package:study_savvy_app/services/jwt_storage.dart';
 abstract class JWTEvent {}
 class JWTEventGet extends JWTEvent{}
 class JWTEventDelete extends JWTEvent{}
+class JWTEventUnknown extends JWTEvent{}
 class JWTBloc extends Bloc<JWTEvent,String?> {
   final JwtService jwtService;
   JWTBloc({JwtService? jwtService}):jwtService=jwtService??JwtService(),super(null){
@@ -12,9 +13,12 @@ class JWTBloc extends Bloc<JWTEvent,String?> {
         String? jwt = await jwtService!.getJwt();
         emit(jwt);
       }
-      else{
+      else if (event is JWTEventDelete){
         await jwtService!.deleteJwt();
         emit(null);
+      }
+      else{
+        throw Exception("Error event in jwt");
       }
     });
   }
